Added information and tools for packaging, changed return values to strings for better ease of use, changed return of model to names instead of "UnknownVariable" objects, unified main.py with main_cmd.py
