+++ pyproject.toml
@@ -1,0 +1,29 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "polyhorn"
+version = "0.0.1"
+authors = [
+  { name="Milad Saadat", email="milad.saadat@sharif.edu" },
+  { name="Ehsan Goharshady", email="ehsan.goharshady@ist.ac.at" },
+  { name="Mehrdad Karrabi", email="mehrdad.karrabi@ist.ac.at" },
+  { name="Maximilian Seeliger", email="maximilian.seeliger@tuwien.ac.at" }
+]
+description = "PolyHorn is a solver for Polynomial Constrained Horn Clauses (pCHC)."
+readme = "README.md"
+requires-python = ">=3.9"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    "numpy",
+    "lark"
+]
+
+[project.urls]
+Homepage = "https://github.com/milad-saadat/PolyHorn"
+Issues = "https://github.com/milad-saadat/PolyHorn/issues"
