+++ src/polyhorn/main.py
@@ -3,8 +3,7 @@ import json
 import uuid
 import string
 import random
-from src.Parser import *
-from src.PositiveModel import Result
+from polyhorn.Parser import *
 
 
 def load_config(config_path: str) -> dict:
@@ -122,3 +121,29 @@ def execute(config_path: str, input: str, parser_method):
         if not output_path_exists:
             os.remove(config["output_path"])
     return sat, model
+
+
+if __name__ == "__main__":
+    from argparse import ArgumentParser
+    parser = ArgumentParser()
+    parser.add_argument("--smt2", type=str, help="Path to the smt2 file")
+    parser.add_argument("--readable", type=str, help="Path to the readable file")
+    parser.add_argument("--config", type=str, required=True, help="Path to the config file")
+    args = parser.parse_args()
+
+    if args.smt2:
+        with open(args.smt2, "r") as file:
+            smt2 = file.read()
+        is_sat, model = execute_smt2(args.config, smt2)
+    elif args.readable:
+        with open(args.readable, "r") as file:
+            readable = file.read()
+        is_sat, model = execute_readable(args.config, readable)
+    else:
+        raise ValueError("Either --smt2 or --readable must be provided")
+    
+    print(f"The system is {is_sat}")
+    if is_sat == 'sat':
+        print("Model:")
+        for var, value in model.items():
+            print(f"{var}: {value}")
